Renamed meta to metadata to ensure it works with std enabled.
