+++ irmaseal-core/src/metadata.rs
